Refactor child stack and navigation out of router
